% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{addPath,reducedDim-method}
\alias{addPath,reducedDim-method}
\title{Store path}
\usage{
\S4method{addPath}{reducedDim}(object, pathName, pathDescription, path)
}
\arguments{
\item{object}{Dimensionality reduced object}

\item{pathName}{Name given to the path}

\item{pathDescription}{A description of the path}

\item{path}{A numeric vector of cells on the path}
}
\value{
Dimensionality reduced object
}
\description{
Given a vector of cells, this will add a path to the paths slot.
}
\details{
Given a vector of cells, this will add a path to the paths slot.
}
\author{
Wajid Jawaid
}

