% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{addPath,SCD-method}
\alias{addPath,SCD-method}
\title{Store path}
\usage{
\S4method{addPath}{SCD}(object, reducedDim, pathName, pathDescription, path,
  ...)
}
\arguments{
\item{object}{Single cell dataset or dimensionality reduced object}

\item{reducedDim}{A dimensionality reduced object}

\item{pathName}{Name given to the path}

\item{pathDescription}{A decription of the path}

\item{path}{A numeric vector of cells on the path}

\item{...}{Parameters passed to the Method \code{addPath}}
}
\value{
Dimensionality reduced object
}
\description{
Given a vector of cells, this will add a path to the paths slot.
}
\details{
Given a vector of cells, this will add a path to the paths slot.
}
\author{
Wajid Jawaid
}

