% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{addPheno,SCD-method}
\alias{addPheno,SCD-method}
\title{Add additional phenotype data}
\usage{
\S4method{addPheno}{SCD}(object, value, noprocessing = FALSE)
}
\arguments{
\item{object}{Single Cell Dataset}

\item{value}{Value of pData, must have same length as the number of cells / samples.}

\item{noprocessing}{Default FALSE. Will return unfiltered/unordered results if set to TRUE}
}
\value{
Updated Single Cell Dataset object
}
\description{
Add additional phenotype data
}
\details{
Add additional phenotype data
}
\author{
Wajid Jawaid
}

