% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{applyMVP}
\alias{applyMVP}
\title{Applies transformation from rgl plot to data}
\usage{
applyMVP(data, MVP = NULL)
}
\arguments{
\item{data}{3D matrix of value to be plotted, with each dimension in a column}

\item{MVP}{Default NULL. The transformation matrix of Projection matrix %*% Viewer matrix %*%
Model matrix. This can be retrieved using the \code{getMVP} function or if not supplied is
retrieved from the current open rgl device.}
}
\value{
Returns the transformed data ready to be plotted using the R base plot function.
}
\description{
Applies transformation from rgl plot to data
}
\details{
Applies transformation from rgl plot to data
}
\author{
Wajid Jawaid
}

