% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{board_ratio}
\alias{board_ratio}
\title{The height/width ratio of diagrams}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
board_ratio
}
\description{
This variable stores the height/width ratio of the diagrams
produced by \code{ggboard()}. It is shared among many helper functions.
}
\keyword{datasets}
