% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{checker_radius}
\alias{checker_radius}
\title{The radius of checkers}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
checker_radius
}
\description{
This variable stores the radius of checkers used in
diagrams produced by \code{ggboard()}. For use in some helper functions.
}
\keyword{datasets}
