% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clipboard}
\alias{clipboard}
\title{Copy R content to clipboard}
\usage{
clipboard(x, sep = "\\n", row.names = FALSE, col.names = TRUE)
}
\arguments{
\item{x}{Vector of characters}

\item{sep}{Default = "\\n". Separator.}

\item{row.names}{Not used.}

\item{col.names}{Not used.}
}
\value{
Copies vector to clipboard.
}
\description{
Copy R content to clipboard
}
\details{
Copy R content to clipboard
}
\author{
Wajid Jawaid
}

