% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{color_schemes}
\alias{color_schemes}
\title{Color schemes}
\format{
An object of class \code{list} of length 2.
}
\usage{
color_schemes
}
\description{
This list stores colors uses in different color schemes.
}
\keyword{datasets}
