% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{counts,SCD-method}
\alias{counts,SCD-method}
\title{Retrieve counts data from Single Cell Dataset object}
\usage{
\S4method{counts}{SCD}(object, ...)
}
\arguments{
\item{object}{SCD}

\item{...}{Additional parameters}
}
\value{
matrix of counts
}
\description{
Retrieve gene counts from Single Cell Dataset object
}
\details{
Retrieve gene counts from Single Cell Dataset object
}
\author{
Wajid Jawaid
}

