% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{diffuse}
\alias{diffuse}
\title{Diffusion Maps}
\usage{
diffuse(object, ndims = 4, nn = 0.2, sigma = NULL, removeFirst = TRUE,
  useARPACK = TRUE, distfun = NULL, ...)
}
\arguments{
\item{object}{Single Cell Data object}

\item{ndims}{Number of dimensions required if using ARPACK otherwise all will be returned}

\item{nn}{Number of nearest neighbours. Can be set to 1}

\item{sigma}{Default NULL. Parameter for Gaussian kernel.}

\item{removeFirst}{Remove the non informative first eigen vector}

\item{useARPACK}{By default uses rARPACK but can be set to false to use the eigen
function from the base package.}

\item{distfun}{A distance function.}

\item{...}{Other parameters to pass on.}
}
\value{
Single Cell Data object with diffusion map
}
\description{
Diffusion map dimensionality reduction
}
\details{
Diffusion map dimensionality reduction as performed by Florian Buettner / Fabian Theis
Diffusion map dimensionality reduction
}
\author{
Wajid Jawaid
}
