% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{diffuseMat}
\alias{diffuseMat}
\title{Generic diffusion function}
\usage{
diffuseMat(data, ndims = 4, nn = 0.2, sigma = 12, removeFirst = TRUE,
  bluntDensityCorrection = FALSE, useARPACK = TRUE, normBy = c("Column",
  "Row", "Original"), distfun = NULL)
}
\arguments{
\item{data}{Matrix of data with genes in rows and cells in columns.}

\item{ndims}{Number of dimensions required if using ARPACK otherwise all will be returned}

\item{nn}{Number of nearest neighbours. Can be set to 1}

\item{sigma}{Default NULL. Parameter for Gaussian kernel.}

\item{removeFirst}{Remove the non informative first eigen vector}

\item{bluntDensityCorrection}{Blunt density correction}

\item{useARPACK}{By default uses rARPACK but can be set to false to use the eigen
function from the base package.}

\item{normBy}{Normalise by Column, row or both.}

\item{distfun}{A different distance function}
}
\value{
List conatining diffusion map.
}
\description{
Generic diffusion function
}
\details{
Generic diffusion function
}
\author{
Wajid Jawaid
}
