% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_defs.R
\docType{class}
\name{DiffusionMap-class}
\alias{DiffusionMap-class}
\title{Diffusion Map dimensionality reduction on dataset}
\description{
Performed as designed by Florian Buettner
}
\section{Slots}{

\describe{
\item{\code{numberConverged}}{The number of converged eigenvalues and eigenvectors when using
"rARPACK"}

\item{\code{numberIterations}}{The number of iterations to achieve convergence in "rARPACK".}

\item{\code{sigmas}}{Value of sigma used for the Gaussian}

\item{\code{nn}}{Number of nearest neighours used}

\item{\code{usedARPACK}}{Logical indicating whether rARPACK was used for eigen deconstruction}

\item{\code{gaussian}}{Post applying Gaussian kernel.}

\item{\code{markov}}{Gaussian with rows normalised giving stochastic transition matrix.}

\item{\code{H}}{H matrix.}

\item{\code{d2}}{Squared ditance matrix.}

\item{\code{distfun}}{Function used to calculate distances.}
}}

