% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{dim,SCD-method}
\alias{dim,SCD-method}
\title{Dimensions of SCD}
\usage{
\S4method{dim}{SCD}(x)
}
\arguments{
\item{x}{Single cell data set}
}
\value{
dimension of object
}
\description{
Dimensions of SCD
}
\details{
Dimensions of SCD
}
\author{
Wajid Jawaid
}

