% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{doDE}
\alias{doDE}
\title{Differential expression using Wilcoxon Ranks sum test}
\usage{
doDE(scd, group1, group2, thr = 1e-04, sdFilter = TRUE,
  colSideCols = NULL, cscCols = "NULL", minCounts = 1)
}
\arguments{
\item{scd}{Single cell data set object}

\item{group1}{Names of cells in group 1}

\item{group2}{Names of cells in group 2}

\item{thr}{Default 1e-4. False discovery rate threshold}

\item{sdFilter}{Default TRUE. Remove genes with zero standard deviation}

\item{colSideCols}{Which column to use for addtional column side colors for heatmap.}

\item{cscCols}{Colours to use for the different levels}

\item{minCounts}{Only include genes with more than this number of counts
over all the cells used in DE.}
}
\value{
Returns a list object
}
\description{
Differential expression
}
\details{
Identify differentially expressed genes using Wilcoxon Rank sum test
}
\author{
Wajid Jawaid
}
