% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{doDESeq}
\alias{doDESeq}
\title{Differential expression using DESeq2}
\usage{
doDESeq(scd, group1, group2, fdr, gene = NULL, rmPattern = NULL,
  minCounts = 1, nMinCells = ceiling(min(length(group1), length(group2))/2))
}
\arguments{
\item{scd}{SCD}

\item{group1}{First group. Query group.}

\item{group2}{Second group. Reference group.}

\item{fdr}{FDR}

\item{gene}{Vector of genes (ENSEMBL ids) to use for DE.}

\item{rmPattern}{Remove these genes}

\item{minCounts}{Remove genes with raw counts less than this}

\item{nMinCells}{The minimum number of cells that must
have the gene expressed (more than 0 counts) for the gene
to be included for differential expression.}
}
\value{
List.
}
\description{
Differential expression using DESeq2
}
\details{
Differential expression using DESeq2
}
\author{
Wajid Jawaid
}

