% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{eigenvals,DiffusionMap-method}
\alias{eigenvals,DiffusionMap-method}
\title{Diffusion Map eigen values}
\usage{
\S4method{eigenvals}{DiffusionMap}(object)
}
\arguments{
\item{object}{DiffusionMap object}
}
\value{
Numeric vector of eigen values for
}
\description{
Returns eigen values for Diffusion Map object
}
\details{
In the interest of time the diffusion map algorithm only calculates
a given number of largest eigen values and vectors using rARPACK.
Therefore not all eigen values will be returned.
Returns eigen values for Diffusion Map object
}
\author{
Wajid Jawaid
}
