% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{eigenvals,reducedDim-method}
\alias{eigenvals,reducedDim-method}
\title{Eigen values of Reduced Dimension object}
\usage{
\S4method{eigenvals}{reducedDim}(object)
}
\arguments{
\item{object}{Dimensionality reduced object}
}
\value{
numeric vestor of ordered eigen values
}
\description{
Returns eigen values
}
\details{
Returns eigen values
}
\author{
Wajid Jawaid
}
