% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{eigenvals}
\alias{eigenvals}
\title{Eigen values}
\usage{
eigenvals(object)
}
\arguments{
\item{object}{Dimensionality reduced object}
}
\value{
numeric vestor of ordered eigen values
}
\description{
Returns eigen values
}
\details{
Returns eigen values
}
\author{
Wajid Jawaid
}
