% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{eigenvecs,DiffusionMap-method}
\alias{eigenvecs,DiffusionMap-method}
\title{Diffusion Map eigen vectors}
\usage{
\S4method{eigenvecs}{DiffusionMap}(object)
}
\arguments{
\item{object}{DiffusionMap object}
}
\value{
Annotated matrix of eigen vectors, with eigen vectors in columns.
}
\description{
Returns eigen vectors for Diffusion Map object
}
\details{
In the interest of time the diffusion map algorithm only calculates
a given number of largest eigen values and vectors using rARPACK.
Therefore not all eigen vectors will be returned.
Returns eigen vectors for Diffusion Map object
}
\author{
Wajid Jawaid
}
