% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{eigenvecs,reducedDim-method}
\alias{eigenvecs,reducedDim-method}
\title{Eigen vectors}
\usage{
\S4method{eigenvecs}{reducedDim}(object)
}
\arguments{
\item{object}{Dimensionality reduced object}
}
\value{
Annotated matrix of eigen vectors, with eigen vectors in columns.
}
\description{
Returns eigen vectors
}
\details{
Returns eigen values
}
\author{
Wajid Jawaid
}
