% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{eigenvecs}
\alias{eigenvecs}
\title{Eigen vectors}
\usage{
eigenvecs(object)
}
\arguments{
\item{object}{Dimensionality reduced object}
}
\value{
Annotated matrix of eigen vectors, with eigen vectors in columns.
}
\description{
Returns eigen vectors
}
\details{
Returns eigen values
}
\author{
Wajid Jawaid
}
