% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_janowski}
\alias{eq_janowski}
\title{Calculate equity for money game, Janowski-style}
\usage{
eq_janowski(outcome_probs, C, p, x = 2/3)
}
\arguments{
\item{outcome_probs}{numeric vector of length 6, representing outcome
probabilities (must always sum to 1 or 100)}

\item{C}{Cube position: 0: Center; 1: player; -1: opponent}

\item{p}{Probability of winning}

\item{x}{cube-life index, between 0 and 1}
}
\value{
double. Equity
}
\description{
Calculate equity for money game, Janowski-style
}
\examples{
probs <- c(31, 4, 0, 47, 17, 1)
eq_janowski(outcome_probs = probs, C = 1, p = 0.5)

}
