% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estSizeFact}
\alias{estSizeFact}
\title{Estimates size factors as in DESeq.}
\usage{
estSizeFact(x)
}
\arguments{
\item{x}{Matrix of values. Cells in columns. Genes in rows.}
}
\value{
Numeric vector of size factors.
}
\description{
Estimates size factors as in DESeq.
}
\details{
Estimates size factors as in DESeq.
}
\author{
Wajid Jawaid
}
