% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estSizeFact2}
\alias{estSizeFact2}
\title{Estimate size factor supercharged}
\usage{
estSizeFact2(x)
}
\arguments{
\item{x}{Matrix of gene expression}
}
\value{
Size factors
}
\description{
Estimate size factor supercharged
}
\details{
Estimate size factor supercharged
}
\author{
wj241
}
