% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{excludeCells,SCD-method}
\alias{excludeCells,SCD-method}
\title{Exclude selected cells}
\usage{
\S4method{excludeCells}{SCD}(object, regexpr = NULL, cellNames = NULL,
  invert = FALSE, reset = TRUE)
}
\arguments{
\item{object}{Single Cell Dataset object}

\item{regexpr}{A regular expression of the cell names to be removed}

\item{cellNames}{A vector of cell names to be removed}

\item{invert}{Default FALSE. If TRUE include the given names and exclude other cells.}

\item{reset}{Default TRUE. If TRUE reset allow choice from all cells in dataset.}
}
\value{
modified Single Cell Data object
}
\description{
Exclude selected cells
}
\details{
Function to remove selected cells from analyses
}
\author{
Wajid Jawaid
}

