% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exportToMatlab}
\alias{exportToMatlab}
\title{Writes data from single cell dataset object for use in Matlab}
\usage{
exportToMatlab(hscd, file, cols, colourBy)
}
\arguments{
\item{hscd}{Single cell dataset object}

\item{file}{Filename (and path) to save file to}

\item{cols}{Vector of colors}

\item{colourBy}{Name of the column in phenotype data by which to colour cells}
}
\value{
Generates a ".mat " file
}
\description{
Writes data from single cell dataset object for use in Matlab
}
\details{
Writes data from single cell dataset object for use in Matlab
}
\author{
Wajid Jawaid
}
