% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expressedGene}
\alias{expressedGene}
\title{Identify expressed genes}
\usage{
expressedGene(x)
}
\arguments{
\item{x}{Gene expression matrix.}
}
\value{
Logical vector indicating whether gene is expressed.
}
\description{
Identify expressed genes
}
\details{
Identify expressed genes
}
\author{
wj241
}

