% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{exprs,SCD-method}
\alias{exprs,SCD-method}
\title{Retrieve gene expression data from Single Cell Dataset object}
\usage{
\S4method{exprs}{SCD}(object)
}
\arguments{
\item{object}{SCD}
}
\value{
matrix of expression values
}
\description{
Retrieve gene expression data from Single Cell Dataset object
}
\details{
Retrieve gene expression data from Single Cell Dataset object
}
\author{
Wajid Jawaid
}
