% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{facsPlot}
\alias{facsPlot}
\title{Generate a FACS like plot from single-cell RNAseq data}
\usage{
facsPlot(scd, gene1, gene2, t1 = NULL, t2 = NULL, filterCol = NULL,
  filterVal = NULL, nrpoints = Inf, pch = 16, nbin = 200,
  bandwidth = 0.1, cex = 0.3, colramp = colorRampPalette(brewer.pal(9,
  "YlOrRd")), t.bold = TRUE, t.lty = 2, t.col = "azure", ...)
}
\arguments{
\item{scd}{Single Cell Dataset}

\item{gene1}{Gene to plot on x-axis}

\item{gene2}{Gene to plot on y-axis}

\item{t1}{threshold 1}

\item{t2}{threshold 2}

\item{filterCol}{perform only on cells with a unique measure on this feature}

\item{filterVal}{Linked to filterVal}

\item{nrpoints}{Deafult Inf.}

\item{pch}{Default 0.3.}

\item{nbin}{Default 200.}

\item{bandwidth}{Default 0.1.}

\item{cex}{Default 0.3.}

\item{colramp}{See smoothScatter() function.}

\item{t.bold}{Default TRUE. To draw bold black lines for the thresholds.}

\item{t.lty}{Default 2. For supplementary threshold line type.}

\item{t.col}{Default "azure". Colour for supplementary threshold line.}

\item{...}{Additional parameters passed to smoothScatter()}
}
\value{
generates plot
}
\description{
Generate a FACS like plot from single-cell RNAseq data
}
\details{
Generate a FACS like plot from single-cell RNAseq data
}
\author{
Wajid Jawaid
}

