% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{fData,SCD-method}
\alias{fData,SCD-method}
\title{Retrieve gene information table from Single Cell Dataset object}
\usage{
\S4method{fData}{SCD}(object)
}
\arguments{
\item{object}{SCD}
}
\value{
Data frame of ENSEMBL ids and corresponding human readable gene names
}
\description{
Retrieve gene information table from Single Cell Dataset object
}
\details{
Retrieve gene information table from Single Cell Dataset object
This table should include the ENSEMBL ids matching the gene names in the gene expression
matrix along with the corresponding human readable gene names.
}
\author{
Wajid Jawaid
}

