% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{filterCell,SCD-method}
\alias{filterCell,SCD-method}
\title{Predicate indicating whether cell filtering is active.}
\usage{
\S4method{filterCell}{SCD}(object)
}
\arguments{
\item{object}{Single Cell Dataset}
}
\value{
Logical
}
\description{
To use filter or not
}
\details{
To use filter or not. Replaces "noprocessing" argument from previous version
}
\author{
Wajid Jawaid
}
