% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{filterCell<-,SCD,logical-method}
\alias{filterCell<-,SCD,logical-method}
\title{Set cell filter.}
\usage{
\S4method{filterCell}{SCD,logical}(object) <- value
}
\arguments{
\item{object}{SCD object}

\item{value}{Logical value.}
}
\value{
Updated SCD object
}
\description{
Set whether to use cell filter or not
}
\details{
Set whether to use cell filter or not
}
\author{
Wajid Jawaid
}

