% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{filterCell<-}
\alias{filterCell<-}
\title{Set whether to use filter or not}
\usage{
filterCell(object) <- value
}
\arguments{
\item{object}{SCD object}

\item{value}{Logical value.}
}
\value{
Updated SCD object
}
\description{
Set whether to use filter or not
}
\details{
Set whether to use filter or not
}
\author{
Wajid Jawaid
}
