% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{filterCell}
\alias{filterCell}
\title{Predicate indicating whether cell filtering is active}
\usage{
filterCell(object)
}
\arguments{
\item{object}{Single Cell Dataset}
}
\value{
Logical.
}
\description{
To use cell filter or not
}
\details{
To use cell filter or not.
}
\author{
Wajid Jawaid
}
