% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{filterGene,SCD-method}
\alias{filterGene,SCD-method}
\title{Predicate indicating whether gene filtering is active}
\usage{
\S4method{filterGene}{SCD}(object)
}
\arguments{
\item{object}{Single Cell Dataset}
}
\value{
Updated Single Cell Dataset object
}
\description{
To use gene filter or not
}
\details{
To use filter or not.
}
\author{
Wajid Jawaid
}

