% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{filterGene<-,SCD,logical-method}
\alias{filterGene<-,SCD,logical-method}
\title{Set gene filter.}
\usage{
\S4method{filterGene}{SCD,logical}(object) <- value
}
\arguments{
\item{object}{SCD object}

\item{value}{Logical value.}
}
\value{
Updated SCD object
}
\description{
Set whether to use gene filter or not
}
\details{
Set whether to use gene filter or not
}
\author{
Wajid Jawaid
}

