% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{filterGene}
\alias{filterGene}
\title{Predicate indicating whether gene filtering is active}
\usage{
filterGene(object)
}
\arguments{
\item{object}{Single Cell Dataset}
}
\value{
Logical.
}
\description{
To use gene filter or not
}
\details{
To use gene filter or not.
}
\author{
Wajid Jawaid
}

