% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{filterQC,SCD-method}
\alias{filterQC,SCD-method}
\title{Predicate indicating whether filtering of cells that failed QC is active}
\usage{
\S4method{filterQC}{SCD}(object)
}
\arguments{
\item{object}{Single Cell Dataset}
}
\value{
Logical
}
\description{
To filter out cells that failed QC.
}
\details{
To filter out cells that failed QC. It is highly recommended that this
is not disabled as size factor and technical noise estimation may fail.
}
\author{
Wajid Jawaid
}
