% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{filterQC<-}
\alias{filterQC<-}
\title{Set whether to use QC filter or not}
\usage{
filterQC(object) <- value
}
\arguments{
\item{object}{SCD object}

\item{value}{Logical value.}
}
\value{
Updated SCD object
}
\description{
Set whether to use QC or not
}
\details{
Set whether to use QC or not
}
\author{
Wajid Jawaid
}

