% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{filterQC}
\alias{filterQC}
\title{Predicate indicating whether filtering of cells that failed QC is active}
\usage{
filterQC(object)
}
\arguments{
\item{object}{Single Cell Dataset}
}
\value{
Logical
}
\description{
To filter out cells that failed QC.
}
\details{
To filter out cells that failed QC. It is highly recommended that this
is not disabled as size factor and technical noise estimation may fail.
}
\author{
Wajid Jawaid
}

