% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findCells}
\alias{findCells}
\title{Finds cells from SCD object matching query term}
\usage{
findCells(scd, pattern, column)
}
\arguments{
\item{scd}{Single Cell Datast object.}

\item{pattern}{Grep pattern match.}

\item{column}{Column of pData to which pattern should be matched.}
}
\value{
Names of identified cells.
}
\description{
Finds cells from SCD object matching query term
}
\details{
Finds cells from SCD object matching query term
}
\author{
wj241
}
