% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findGene}
\alias{findGene}
\title{Find an approximate gene name}
\usage{
findGene(scd, g, includeHidden = TRUE)
}
\arguments{
\item{scd}{Single Cell data set object}

\item{g}{approximate gene name or alias}

\item{includeHidden}{Search genes that are otherwise excluded.}
}
\value{
returns names from loaded ENSMBL gene table
}
\description{
Find gene name
}
\details{
Find gene name
}
\author{
Wajid Jawaid
}

