% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{findPath,SCD-method}
\alias{findPath,SCD-method}
\title{Find path between two selected cells}
\usage{
\S4method{findPath}{SCD}(object, reducedDim, from, to, pathName,
  pathDescription)
}
\arguments{
\item{object}{A single cell dataset object}

\item{reducedDim}{Reduced diensionality object}

\item{from}{Numeric value representing the starting cell of path}

\item{to}{Numeric value representing the ending cell of path}

\item{pathName}{Name to give path}

\item{pathDescription}{Description of path}
}
\value{
Altered object of same class
}
\description{
Finds and saves path between two cells
}
\details{
Finds and saves path between two cells
}
\author{
Wajid Jawaid
}
