% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{findPath}
\alias{findPath}
\title{Find path between two selected cells}
\usage{
findPath(object, reducedDim, from, to, pathName, pathDescription)
}
\arguments{
\item{object}{A single cell dataset object or a reduced dimensionality object}

\item{reducedDim}{Reduced diensionality object}

\item{from}{Numeric value representing the starting cell of path}

\item{to}{Numeric value representing the ending cell of path}

\item{pathName}{Name to give path}

\item{pathDescription}{Description of path}
}
\value{
Altered object of same class
}
\description{
Finds and saves path between two cells
}
\details{
Finds and saves path between two cells
}
\author{
Wajid Jawaid
}
