% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gammon_value}
\alias{gammon_value}
\title{Calculate gammon value for a score and cube level:
The gain from winning a gammon instead of single game, divided
by the loss from losing a single game instead of winning one.}
\usage{
gammon_value(x, y, cube, met)
}
\arguments{
\item{x}{number of points that player needs}

\item{y}{number of points that opponent needs}

\item{cube}{cube value}

\item{met}{match equity table}
}
\value{
a real number, the value of winning a gammon
}
\description{
Calculate gammon value for a score and cube level:
The gain from winning a gammon instead of single game, divided
by the loss from losing a single game instead of winning one.
}
