% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geneDistribution}
\alias{geneDistribution}
\title{Histogram of gene distribution}
\usage{
geneDistribution(scd, gene, fun = function(x) 10^x, ...)
}
\arguments{
\item{scd}{Single Cell Dataset object}

\item{gene}{Gene to be plotted}

\item{fun}{Function to use on gene expression values. Defaults to log10 normalisedcounts i.e. \eqn{10^{count}}}

\item{...}{Further parameters to pass to \code{hist()}}
}
\value{
Gneerates a plot and returns value from \code{hist()} invisibly.
}
\description{
Plots histogram of gene expression distribution
}
\details{
Plots histogram of gene expression distribution
}
\author{
Wajid Jawaid
}
