% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geneID2Name}
\alias{geneID2Name}
\title{Returns human readable gene names}
\usage{
geneID2Name(scd, genes)
}
\arguments{
\item{scd}{Single cell Dataset object}

\item{genes}{Ensembl gene IDs}
}
\value{
Named vector
}
\description{
Returns human readable gene names
}
\details{
Returns human readable gene names
}
\author{
Wajid Jawaid
}

