% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geneName2id}
\alias{geneName2id}
\title{Returns Ensembl id given a gene name}
\usage{
geneName2id(scd, genes, useAgrep = FALSE)
}
\arguments{
\item{scd}{Single Cell Dataset}

\item{genes}{Gene names}

\item{useAgrep}{To use approximate match}
}
\value{
Named vector
}
\description{
Returns Ensembl id given a gene name
}
\details{
Returns Ensembl id given a gene name
}
\author{
Wajid Jawaid
}

