% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{getDiffMap,SCD-method}
\alias{getDiffMap,SCD-method}
\title{Retrieves diffusion map object}
\usage{
\S4method{getDiffMap}{SCD}(object, ...)
}
\arguments{
\item{object}{Single Cell Dataset object}

\item{...}{Additonal parameters}
}
\value{
Returns DiffMap dimensionality reduced object
}
\description{
Retrieve diffusion map object
}
\details{
Retrieve diffusion map object
}
\author{
Wajid Jawaid
}
