% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getDiffMap}
\alias{getDiffMap}
\title{Retrieves diffusion map object}
\usage{
getDiffMap(object, ...)
}
\arguments{
\item{object}{Single Cell Dataset object}

\item{...}{Additonal parameters}
}
\value{
Returns DiffMap dimensionality reduced object
}
\description{
Retrieve diffusion map object
}
\details{
Retrieve diffusion map object
}
\author{
Wajid Jawaid
}

