% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{getIsomap,SCD-method}
\alias{getIsomap,SCD-method}
\title{Retrieves Isomap object}
\usage{
\S4method{getIsomap}{SCD}(object, ...)
}
\arguments{
\item{object}{Single Cell Dataset object}

\item{...}{Additional parameters}
}
\value{
Returns Isomap dimensionality reduced object
}
\description{
Retrieve Isomap object
}
\details{
Retrieve Isomap object
}
\author{
Wajid Jawaid
}

