% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getIsomap}
\alias{getIsomap}
\title{Retrieves Isomap object}
\usage{
getIsomap(object, ...)
}
\arguments{
\item{object}{Single Cell Dataset object}

\item{...}{Additional parameters}
}
\value{
Returns Isomap dimensionality reduced object
}
\description{
Retrieve Isomap object
}
\details{
Retrieve Isomap object
}
\author{
Wajid Jawaid
}
