% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{getMST,reducedDim-method}
\alias{getMST,reducedDim-method}
\title{Retrieves the minimum spanning tree from the \code{graph} slot of a
Dimensionality reduced object}
\usage{
\S4method{getMST}{reducedDim}(object)
}
\arguments{
\item{object}{Dimensionality reduced object}
}
\value{
Minimum spanning tree as an igraph object
}
\description{
Returns the minimum spanning tree
}
\details{
Returns the minimum spanning tree
}
\author{
Wajid Jawaid
}
