% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getMST}
\alias{getMST}
\title{Retrieves the minimum spanning tree from the \code{graph} slot of a Single
Cell Dataset or a dimensionality reduced object}
\usage{
getMST(object, reducedDim)
}
\arguments{
\item{object}{Single Cell Dataset object or dimensionlaity reduced object}

\item{reducedDim}{The dimensionality reduced object}
}
\value{
Minimum spanning tree as an igraph object
}
\description{
Returns the minimum spanning tree
}
\details{
Returns the minimum spanning tree
}
\author{
Wajid Jawaid
}

