% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getMVP}
\alias{getMVP}
\title{Retrieve rendering data from current rgl plot}
\usage{
getMVP()
}
\value{
Matrix to achieve transformation from model space to the screen
}
\description{
Get rendering data from an open rgl plot
}
\details{
Get rendering data from an open rgl plot
}
\author{
Wajid Jawaid
}
