% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{getPaths,reducedDim-method}
\alias{getPaths,reducedDim-method}
\title{Retrieve paths from a dimensionality reduced object}
\usage{
\S4method{getPaths}{reducedDim}(object)
}
\arguments{
\item{object}{Dimensionality reduced object}
}
\value{
The stored paths for the object
}
\description{
Retrieve paths from a dimensionality reduced object
}
\details{
Retrieve paths from a dimensionality reduced object
}
\author{
Wajid Jawaid
}
