% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getPaths}
\alias{getPaths}
\title{Retrieve paths from a dimensionality reduced object}
\usage{
getPaths(object)
}
\arguments{
\item{object}{Dimensionality reduced object}
}
\value{
The stored paths for the object
}
\description{
Retrieve paths from a dimensionality reduced object
}
\details{
Retrieve paths from a dimensionality reduced object
}
\author{
Wajid Jawaid
}

