% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{getPCA,SCD-method}
\alias{getPCA,SCD-method}
\title{Retrieves PCA object}
\usage{
\S4method{getPCA}{SCD}(object, ...)
}
\arguments{
\item{object}{Single Cell Dataset object}

\item{...}{Additional parameters}
}
\value{
Returns PCA dimensionality reduced object
}
\description{
Retrieve PCA data
}
\details{
Retrieve PCA data
}
\author{
Wajid Jawaid
}
