% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getPCA}
\alias{getPCA}
\title{Retrieves PCA object}
\usage{
getPCA(object, ...)
}
\arguments{
\item{object}{Single Cell Dataset object}

\item{...}{Additional parameters}
}
\value{
Returns PCA dimensionality reduced object
}
\description{
Retrieve PCA data
}
\details{
Retrieve PCA data
}
\author{
Wajid Jawaid
}

