% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{getTSNE,SCD-method}
\alias{getTSNE,SCD-method}
\title{Retrieves t-SNE object}
\usage{
\S4method{getTSNE}{SCD}(object, ...)
}
\arguments{
\item{object}{Single Cell Dataset object}

\item{...}{Additional parameters}
}
\value{
Returns t-SNE dimensionality reduced object
}
\description{
Retrieve t-SNE object
}
\details{
Retrieve t-SNE object
}
\author{
Wajid Jawaid
}
