% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getTSNE}
\alias{getTSNE}
\title{Retrieves t-SNE object}
\usage{
getTSNE(object, ...)
}
\arguments{
\item{object}{Single Cell Dataset object}

\item{...}{Additional parameters}
}
\value{
Returns t-SNE dimensionality reduced object
}
\description{
Retrieve t-SNE object
}
\details{
Retrieve t-SNE object
}
\author{
Wajid Jawaid
}
