% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ggCol}
\alias{ggCol}
\title{Generate ggplot colors}
\usage{
ggCol(x, alpha = 1, returnFull = TRUE, remix = TRUE)
}
\arguments{
\item{x}{Vector of group allocation}

\item{alpha}{Default 1. Measure of opacity.}

\item{returnFull}{Default TRUE. Return value for each vector entry.}

\item{remix}{Mix the colours so sequential colours are not identical.}
}
\value{
Returns colors for each entry of the vector or for each group allocation.
}
\description{
Generate ggplot colors
}
\details{
Generate ggplot colors
}
\author{
Wajid Jawaid
}
