% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{gnu_chars}
\alias{gnu_chars}
\title{Characters used in position and match ids}
\format{
An object of class \code{character} of length 64.
}
\usage{
gnu_chars
}
\description{
This variable stores the set of charactgers used by
GNU Backgammon in position id and match id
}
\keyword{datasets}
