% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap.gottgens}
\alias{heatmap.gottgens}
\title{Heatmap helper function for leaf re-ordering}
\usage{
heatmap.gottgens(x, cor.method = "spearman", clust.method = "ward.D2",
  distfun = NULL, col = NULL, breaks = NULL, reorder = c("both", "row",
  "column", "none"), Rowv = NULL, Colv = NULL, scale = c("none", "row",
  "column"), trace = "none", key.cex = 0.4, ColSideColAxis = 4,
  symbreaks = FALSE, margins = c(3, 3), cexRow = 0.6, keysize = 0.5,
  key.title = "Key", key.par = list(mar = c(2.5, 0.5, 1.5, 0.5)),
  density.info = "none", ...)
}
\arguments{
\item{x}{Matrix}

\item{cor.method}{method to be used to calculate distance. Default is (1 - spearman rank correlation). This is overridden by \code{distfun}.}

\item{clust.method}{clustering method to be used. Default is \code{ward.D2}}

\item{distfun}{function to use for distance calculation.}

\item{col}{define colour. If defines please ensure that \code{breaks} are passed to \code{\link{heatmap.minus}}.}

\item{breaks}{colour breaks.}

\item{reorder}{Logical whether to reorder leaves}

\item{Rowv}{determines if and how the row dendrogram should be reordered.    By default, it is TRUE, which implies dendrogram is computed and reordered based on row means. If NULL or FALSE, then no dendrogram is computed and no reordering is done. If a dendrogram, then it is used "as-is", ie without any reordering. If a vector of integers, then dendrogram is computed and reordered based on the order of the vector.}

\item{Colv}{determines if and how the column dendrogram should be reordered.    Has the options as the Rowv argument above and additionally when x is a square matrix, Colv = \code{Rowv} means that columns should be treated identically to the rows.}

\item{scale}{Scale by "row", "column" or "none"}

\item{trace}{character string indicating whether a solid "trace" line should be drawn across 'row's or down 'column's, 'both' or 'none'. The distance of the line from the center of each color-cell is proportional to the size of the measurement. Defaults to 'column'.}

\item{key.cex}{size of key text}

\item{ColSideColAxis}{changes axis on which the group names are printed. Use \code{2} for left and \code{4} for right.}

\item{symbreaks}{Boolean indicating whether breaks should be made symmetric about 0. Defaults to TRUE if the data includes negative values, and to FALSE otherwise.}

\item{margins}{numeric vector of length 2 containing the margins (see \code{\link{par}}(mar= *)) for column and row names, respectively.}

\item{cexRow}{positive numbers, used as cex.axis in for the row or column axis labeling. The defaults currently only use number of rows or columns, respectively.}

\item{keysize}{numeric value indicating the size of the key}

\item{key.title}{character string for the key title}

\item{key.par}{defines the key's parameters e.g. margins}

\item{density.info}{character string indicating whether to superimpose a 'histogram', a 'density' plot, or no plot ('none') on the color-key.}

\item{...}{Parameters passed to \code{\link{heatmap.minus}}}
}
\value{
Calls \code{\link{heatmap.minus}} and returns list with re-ordering
}
\description{
Heatmap helper function for leaf re-ordering
}
\details{
Heatmap helper function for leaf re-ordering, the Gottgens way.
}
\author{
Wajid Jawaid
}

