% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{initialize,SCD-method}
\alias{initialize,SCD-method}
\title{To initialize a SCD object}
\usage{
\S4method{initialize}{SCD}(.Object, experimentType, assayData, genoData,
  phenoData, spike = NULL, qc = NULL, ...)
}
\arguments{
\item{.Object}{Single Cell Dataset object to be created}

\item{experimentType}{Either "RNAseq" or "qPCR"}

\item{assayData}{Annotated matrix of gene expression data}

\item{genoData}{Annotated data frame of gene information. One column must match rownames
in \code{assayData} the gene expression annotated matrix.}

\item{phenoData}{Annotated data frame of phenotype information. One column must match the colnames in \code{assayData} the gene expression annotated matrix.}

\item{spike}{Default NULL. Include spike-in data e.g. ERCC-92.}

\item{qc}{Default NULL. Include quality control counts from HTseq for example.}

\item{...}{Parameters passed to other Methods}
}
\value{
Returns a newly initialised SCD object
}
\description{
Initialise a SCD object
}
\details{
Initialise a SCD object
}
\author{
Wajid Jawaid
}

