% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_defs.R
\docType{class}
\name{Isomap-class}
\alias{Isomap-class}
\title{Isomap dimensionality reduced object}
\description{
Performed in Matlab
}
\section{Slots}{

\describe{
\item{\code{nn}}{Number of nearest neighbours used in the algorithm}

\item{\code{nnMat}}{Adjacency matrix of nearest neighbours}
}}

