% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{loadings,PCA-method}
\alias{loadings,PCA-method}
\title{PCA loadings}
\usage{
\S4method{loadings}{PCA}(x)
}
\arguments{
\item{x}{Single Cell Dataset object}
}
\value{
Annotated matrix of loadings
}
\description{
Returns loadings from PCA object
}
\details{
Returns loadings from PCA object
}
\author{
Wajid Jawaid
}
