% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logVarGenes}
\alias{logVarGenes}
\title{Find highly variable genes (log linear model)}
\usage{
logVarGenes(scd, minMean = 0, fraction = 0.05, lower = FALSE,
  residualsInLogSpace = TRUE)
}
\arguments{
\item{scd}{Single Cell Dataset}

\item{minMean}{Fit model to genes with mean expression gretaer than minMean}

\item{fraction}{The fraction of cells on}

\item{lower}{By default this function will only winsorise the upper end.
If you also require the bottom end to be winsorised set this to TRUE.}

\item{residualsInLogSpace}{Use lm instead of nls to optimise residuals in log-space.}
}
\value{
Returns a list of means, cv2, fit coefficients and variable genes
}
\description{
Find highly variable genes (log linear model)
}
\details{
Find highly variable genes (log linear model).
Model is fitted to:
\deqn{\frac{\sigma^{2}}{\mu^{2}} = a * \mu^{k}}
Using the non-linear squares method.
}
\author{
Wajid Jawaid
}

