% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{markFailedQC}
\alias{markFailedQC}
\title{Mark cells that have failed QC}
\usage{
markFailedQC(object, cellNames, reset = TRUE)
}
\arguments{
\item{object}{Single Cell Dataset}

\item{cellNames}{Character vector of cell names}

\item{reset}{Default TRUE. Sets all others as PASSED.}
}
\value{
Updated Single Cell Data object.
}
\description{
Mark cells that have failed QC
}
\details{
Mark cells that have failed QC
}
\author{
wj241
}
