% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{markGOI}
\alias{markGOI}
\title{Alters matrix rownames to only include genes of interest}
\usage{
markGOI(scd, data, genes)
}
\arguments{
\item{scd}{SCD object to use for translating gene names}

\item{data}{Matrix with genes in rows}

\item{genes}{vector of genes of interest}
}
\value{
Matrix with only genes of interest in rownames, other genes contain empty string.
}
\description{
Alters matrix rownames to only include genes of interest
}
\details{
Alters matrix rownames to only include genes of interest
}
\author{
wj241
}
