% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{memls}
\alias{memls}
\title{List variables ls() on steroids}
\usage{
memls()
}
\value{
List of variables and memory usage
}
\description{
List variables ls() on steroids
}
\details{
List variables ls() on steroids
}
\author{
wj241
}

