% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{mst,reducedDim-method}
\alias{mst,reducedDim-method}
\title{Minimum Spanning Tree for Reduced Dimensionality Object}
\usage{
\S4method{mst}{reducedDim}(object, weighted = TRUE, useDims = NULL)
}
\arguments{
\item{object}{Dimensionality reduced object}

\item{weighted}{Default TRUE. Passed to graph.adjacency() function in igraph.}

\item{useDims}{Default is NULL.}
}
\value{
Dimensionality reduced object
}
\description{
Generates minmium spanning tree
}
\details{
Generates minmium spanning tree and adds diameter path to \code{paths} slot
}
\author{
Wajid Jawaid
}

