% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{mst,SCD-method}
\alias{mst,SCD-method}
\title{Generate minimum spanning true}
\usage{
\S4method{mst}{SCD}(object, wR, weighted = TRUE, useDims = NULL)
}
\arguments{
\item{object}{Single Cell Dataset object}

\item{wR}{Name of slot containing Dimensionality Reduced object}

\item{weighted}{Default TRUE. Passed to igraph.}

\item{useDims}{Deafult is NULL. Sets dimensions to use.}
}
\value{
Returns single cell data object
}
\description{
Generate minimum spanning true
}
\details{
General minimum spanning true
}
\author{
Wajid Jawaid
}

