% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{mst}
\alias{mst}
\title{Minimum Spanning Tree for Reduced Dimensionality Object}
\usage{
mst(object, ...)
}
\arguments{
\item{object}{Dimensionality reduced object}

\item{...}{Any additional parameters}
}
\value{
Dimensionality reduced object
}
\description{
Generates minmium spanning tree
}
\details{
Generates minmium spanning tree and adds diameter path to \code{paths} slot
}
\author{
Wajid Jawaid
}

