% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiGenes}
\alias{multiGenes}
\title{Generates multiple 3D plots with marked gene expression}
\usage{
multiGenes(scd, reduceMethod, geneVect, nr, nc, res = c(1439, 893),
  useDims = 1:3, height = 7, width = 7, mar = c(0, 0, 3.1, 0),
  pdf.name = NULL, includeFiltered = FALSE, vp = NULL, arrange = TRUE,
  project2D = NULL, layoutMat = NULL, legend = FALSE, ...)
}
\arguments{
\item{scd}{Single cell dataset}

\item{reduceMethod}{dimensionality reduction method to be plotted}

\item{geneVect}{character vector of gene names}

\item{nr}{Number of rows on scree}

\item{nc}{Number of columns on screen}

\item{res}{Default 1439 x 893. Resolution of graphics device. CIMR 3200x1024}

\item{useDims}{Dimensions to be plotted}

\item{height}{height of display windows. Deafaults to 7.}

\item{width}{width of display window. Defaults to 7.}

\item{mar}{Plot margins. Defaults to c(0,0,3.1,0).}

\item{pdf.name}{To output to pdf file. Provide name of file. Defaults to NULL}

\item{includeFiltered}{Defaults to FALSE. To include filtered values.}

\item{vp}{Default NULL. Pass par() or par3d() options to set desired viewpoint.
Particularly useful when generating pdf(s).}

\item{arrange}{Default TRUE. To arrange the windows}

\item{project2D}{Default NULL. Projects 3D graph into 2D by the given transformation
matrix. This matrix must have been produced using \code{getMVP} on an open rgl device.}

\item{layoutMat}{Default NULL. Provides greater control of the layout matrix.}

\item{legend}{Default FALSE. To plot legend or not.}

\item{...}{Parameters to pass to plotGene.}
}
\value{
Plots 3D plots.
}
\description{
Generates multiple 3D plots with marked gene expression
}
\details{
Generates multiple 3D plots with marked gene expression
}
\author{
Wajid Jawaid
}
