% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mwc}
\alias{mwc}
\title{Look up match winning chances at different scores}
\usage{
mwc(a, b, met)
}
\arguments{
\item{a}{number of points that player needs}

\item{b}{number of points that opponent needs}

\item{met}{match equity table}
}
\value{
double. Match winning chance
}
\description{
Look up match winning chances at different scores
}
\examples{
met <- get_met() # Get the default Kazaross XG2 table
mwc(3, 5, met)   # Match winning chance at 3-away, 5-away

}
