% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{neatLabels}
\alias{neatLabels}
\title{Decorate plots with log10 values}
\usage{
neatLabels()
}
\value{
Values added to plot
}
\description{
Decorate plots with log10 values
}
\details{
Decorate plots with log10 values. The plot must have prviously been generated.
}
\author{
Wajid Jawaid
}
