% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{newSCD}
\alias{newSCD}
\title{Make a new SCD object}
\usage{
newSCD(experimentType, assayData = NULL, genoData, phenoData, counts,
  spike = NULL, qc = NULL, ...)
}
\arguments{
\item{experimentType}{Experiment Type. RNAseq or qPCR.}

\item{assayData}{Matrix of data with genes in rows and cells in columns}

\item{genoData}{Data frame of gene names with geneIds in rownames. Rownames must match rownames of assayData.}

\item{phenoData}{Data frame of further details for each cell. Rownames must match column names of assayData.}

\item{counts}{Raw counts data.}

\item{spike}{Default NULL. Include spike-in data e.g. ERCC-92.}

\item{qc}{Default NULL. Include quality control counts from HTseq for example.}

\item{...}{Further parameters.}
}
\value{
New Single Cell Dataset object.
}
\description{
Make new SCD object
}
\details{
Construct a new Single cell dataset object for the bglab package
}
\author{
Wajid Jawaid
}

