% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{normalize,SCD-method}
\alias{normalize,SCD-method}
\title{Normalise counts data}
\usage{
\S4method{normalize}{SCD}(object, normMethod = c("DESeq", "scran"),
  logFnc = log10, pseudocount = 1, ...)
}
\arguments{
\item{object}{SCD}

\item{normMethod}{DESeq or scran}

\item{logFnc}{log10}

\item{pseudocount}{Default 1. Added to normalised counts before log transformation}

\item{...}{Additional parameters}
}
\value{
SCD object
}
\description{
Normalise counts data
}
\details{
Normalise counts data
}
\author{
Wajid Jawaid
}
