% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{orderFrame}
\alias{orderFrame}
\title{Reorder frame}
\usage{
orderFrame(object)
}
\arguments{
\item{object}{Annotated data frame}
}
\value{
Reordered frame
}
\description{
Reorder data frame
}
\details{
Reorder data frame
}
\author{
Wajid Jawaid
}
