% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{outcome_probs}
\alias{outcome_probs}
\title{Convert outcome distributions from eXtreme Gammon to probabilities}
\usage{
outcome_probs(xg_probs)
}
\arguments{
\item{xg_probs}{numeric vector of length 6, corresponding to the
winning chances reported by eXtreme Gammon. Can be percentages or
decimal fractions.}
}
\value{
named numeric vector of length 6, representing outcome
probabilities (always sum to 1 or 100)
}
\description{
Convert outcome distributions from eXtreme Gammon to probabilities
}
\examples{
# XGID=-a-BaBC-A---eE---c-e----B-:0:0:1:00:0:0:0:0:10
# 4-ply winning chances:
outcome_probs(c(61.94, 24.09, 1.04, 38.06, 8.54, 0.42))

}
