% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_defs.R
\docType{class}
\name{PCA-class}
\alias{PCA-class}
\title{Princincple Component Analysis on dataset}
\description{
Contains the PCA reduced data. Inheroted from the virtual class \linkS4class{reducedDim}
}
\section{Slots}{

\describe{
\item{\code{standardDeviation}}{Contains the calculated standard deviations}

\item{\code{rotation}}{The transformation to rotate the data}

\item{\code{center}}{Returned from \code{\link{prcomp}}}

\item{\code{scaled}}{Logical value stores whether the data was scaled prior to PCA}
}}

