% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{pData,SCD-method}
\alias{pData,SCD-method}
\title{Retrieve phenotype data}
\usage{
\S4method{pData}{SCD}(object)
}
\arguments{
\item{object}{SCD}
}
\value{
Data frame of phenotype data
}
\description{
Retrieve phenotype information table from Single Cell Dataset object
}
\details{
Retrieve phenotype information table from Single Cell Dataset object
This table should include the sample/cell ID, sequencing chip and index tag, embryo stage
of the sample, the embryo ID and other data - for now including FACS index data
}
\author{
Wajid Jawaid
}

