% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{performQC}
\alias{performQC}
\title{Perform QC on data}
\usage{
performQC(object, selectedCells = "ALL", cutoffs = c(2e+05, 0.2, 0.2, 0, 0,
  1, 0), metaLaneID = "flowCell", mitochondrialIdenitifier = "^mt|^MT",
  pdf = NULL, qcFeatures = "ALL")
}
\arguments{
\item{object}{SCD}

\item{selectedCells}{Default "ALL". If data not available on all cells give subset.}

\item{cutoffs}{Default 2e5, 0.2, 0.2, 0, 0, 1, 0.
Set thresholds for:
\itemize{
 \item{number of mapped reads (greater than threshold), }
 \item{ratio of genes to total number of reads (greater than threshold), }
 \item{ratio of mitchondrial genes to mitochondrial + nuclear genes (less than threshold),}
 \item{number of nuclear genes (greater than threshold), }
 \item{number of genes with >10 reads per million (greater than threshold),}
 \item{ercc:mapped (less than threshold) and}
 \item{nuclearGenes:mapped (greater than threshold).}
}}

\item{metaLaneID}{ID of column that contains information about flow cell lane.}

\item{mitochondrialIdenitifier}{Regexp identifying mitochondrial gene in geneTable}

\item{pdf}{Prefix name of pdfs.}

\item{qcFeatures}{Features from the QC to plot.}
}
\value{
SCD object
}
\description{
Perform QC on data
}
\details{
Perform QC on data
}
\author{
Wajid Jawaid
}

