% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plot,SCD,ANY-method}
\alias{plot,SCD,ANY-method}
\title{General plot function}
\usage{
\S4method{plot}{SCD,ANY}(x, reduceMethod = c("pca", "diffMap", "tsne",
  "isomap"), dims = c(2, 3), colorBy = "embryoStage", useDims = NULL,
  logContinuous = TRUE, opacity = NULL, ...)
}
\arguments{
\item{x}{Single Cell Dataset object}

\item{reduceMethod}{Dimensionality reduction method to choose. At the moment can be either "pca"
or "diffMap"}

\item{dims}{Number of dimensions}

\item{colorBy}{Which column of the annotated phenotype data frame to use for colouring points}

\item{useDims}{Dimensions to be plotted}

\item{logContinuous}{Take log for continuous variables}

\item{opacity}{Default is NULL. Set opacity for each point in pData and pass column
name as character vector. Values in the range 0:255.}

\item{...}{Parameters passed to \code{plot3d} or \code{plot}}
}
\value{
Returns the requested plot
}
\description{
General Plot function
}
\details{
Plot function
}
\author{
Wajid Jawaid
}

