% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot3}
\alias{plot3}
\title{3D plotting function}
\usage{
plot3(object, colorBy = "", plotCols = NULL, legendSize = 1,
  legendOffset = 0.2, legPlac = 1:3, legPos = "front", negLeg = FALSE,
  plotLegend = TRUE, legVert = 1, legTextPos = 1, axLabs = "Component",
  doPath = FALSE, doTree = FALSE, selectedPath = NULL, useDims = 1:3,
  selectedCells = c(NA, NULL), legendTitle = "", diamWidth = 2,
  diamCol = "black", ...)
}
\arguments{
\item{object}{A Reduced Dimension object}

\item{colorBy}{Character or numeric vector. Character vectors will be converted to a factor
and cells coloured accordingly. Cell color can be supplied in the plotCols parameter. If not
supplied the ggplots2 coloring scheme will be adopted.}

\item{plotCols}{Colors to use for the plot}

\item{legendSize}{Controls the size of the legend}

\item{legendOffset}{Distance from the edge of the axis to place the legend as a proportion
of the length of the axis.}

\item{legPlac}{Controls the axis along which the legend is placed.}

\item{legPos}{Controls the depth at which the legend is placed. Default is "front".}

\item{negLeg}{Moves the legend to the other extreme of the axis}

\item{plotLegend}{Logical value. Default is "True"}

\item{legVert}{Controls the vertical orientation of the legend}

\item{legTextPos}{Controls the position of the legend text}

\item{axLabs}{Prefix with which to label the axes.}

\item{doPath}{Logical value controlling whether the diameter path is plotted}

\item{doTree}{Logical calue controlling whether the minimum spanning tree is plotted}

\item{selectedPath}{The path to plot. Default is NULL and will plot the diameter path.}

\item{useDims}{The dimensions of the the dimensionality reduced objects to use}

\item{selectedCells}{The selected cells to plot}

\item{legendTitle}{Legend title.}

\item{diamWidth}{Width of diameter path line}

\item{diamCol}{Color of diameter path line}

\item{...}{Parameters to be passed to \code{plot3d} function.}
}
\value{
3D plot
}
\description{
3D plotting function
}
\details{
3D plotting function
}
\author{
Wajid Jawaid
}

