% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotGene}
\alias{plotGene}
\title{Visual representation of gene expression.}
\usage{
plotGene(scd, gene, useDims = 1:2, reduceMethod = c("pca", "diffMap",
  "tsne", "isomap"), lcol = "grey", hcol = "red", use.ggPlot = FALSE,
  main = gene, pch = 20, includeHidden = FALSE, bringToFront = TRUE,
  project2D = NULL, opacity = NULL, legPos = "topright",
  empty.channel = c(0.2, 0.2, 0.2, 0.5), adjust = 0, minVal = NULL,
  maxVal = NULL, legend = FALSE, multigene = FALSE, useMeta = FALSE,
  bgCol = "black", leg.text.col = "white", leg.inset = 0,
  plt.mar = rep(0, 4), channelOrder = 1:3, ...)
}
\arguments{
\item{scd}{single cell dataset object}

\item{gene}{Gene to be visualised.}

\item{useDims}{dimensions to be used}

\item{reduceMethod}{The dimensionality reduction technique to be used}

\item{lcol}{Colour to be taken for low values. Default is "grey".}

\item{hcol}{Colour to be taken by high values. Default is "red".}

\item{use.ggPlot}{Plot using ggPlots or native R graphics.}

\item{main}{Plot title. Defaults to gene name.}

\item{pch}{Point caharcter. Default is 16.}

\item{includeHidden}{Default FALSE, will only include highly variable genes.}

\item{bringToFront}{Bring high values to fore.}

\item{project2D}{Default NULL. Projects 3D graph into 2D by the given transformation
matrix. This matrix must have been produced using \code{getMVP} on an open rgl device.}

\item{opacity}{Default NULL. Set opacity.}

\item{legPos}{Default "topright". Position of legend}

\item{empty.channel}{Default (0.2,0.2,0.2). Sets the zero state of cell
not expressing anything}

\item{adjust}{Default (0,0,0). Adjustment for red, green and blue channel,
respectively.}

\item{minVal}{Default NULL. Set the maximum and minimum color values manually.}

\item{maxVal}{Default NULL. Set the maximum and minimum color values manually.}

\item{legend}{Default FALSE. Plot legend or not.}

\item{multigene}{Default FALSE. For internal use indicating function
is being called from the multiGene() function.}

\item{useMeta}{Default FALSE. To use information from the meta data.}

\item{bgCol}{Default black. Background colour.}

\item{leg.text.col}{Default white. Legend text colour.}

\item{leg.inset}{Default 0. Legend inset.}

\item{plt.mar}{Default rep(0,4). Plot margins}

\item{channelOrder}{Default 1:3 indicating red, green, blue.}

\item{...}{Additional parameters to pass to plot}
}
\value{
Produced plot
}
\description{
Plots dimensionality reduced object with points coloured by level of gene expression.
}
\details{
Plots dimensionality reduced object with points coloured by level of gene expression.
}
\author{
Wajid Jawaid
}
