% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotGeneProfile}
\alias{plotGeneProfile}
\title{Plots gene profile on trajectory}
\usage{
plotGeneProfile(scd, gene, reduceMethod = "diffMap", paths,
  pathlty = 1:length(paths), geneCol = 1:length(gene), winLength = 10,
  legendPos = "topright", tpos = 0, useGam = TRUE, ...)
}
\arguments{
\item{scd}{SCD object}

\item{gene}{Character vector naming gene of interest}

\item{reduceMethod}{Default "diffMap". Dimensionality reduction method.}

\item{paths}{Vector of path names. Path names for which gene profiles are to be plotted}

\item{pathlty}{Vector of line types for different paths. Give in same order as paths.}

\item{geneCol}{Vector of colors. Give in same order as genes.}

\item{winLength}{Integer. Window length to use for sliding average}

\item{legendPos}{Character vector of length 1. Giving position of legend or
optionally "none" if no legend is to be plotted.}

\item{tpos}{Default 0. Position where to place the title.}

\item{useGam}{Use GAM for smoothing}

\item{...}{Additional parameter to pass to plot.}
}
\value{
Returns plot and a list of results.
}
\description{
Plots gene profile on trajectory
}
\details{
Plots gene profile on trajectory
}
\author{
Wajid Jawaid
}
