% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotLegend}
\alias{plotLegend}
\title{Plots legend for plotGene() function}
\usage{
plotLegend(x, mpar = list(mar = c(4, 4, 2, 2)))
}
\arguments{
\item{x}{output from plotgene}

\item{mpar}{plot parameters}
}
\value{
generates plot
}
\description{
Plots legend for plotGene() function
}
\details{
Plots legend for plotGene() function when using multiple genes
}
\author{
Wajid Jawaid
}

