% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plotPath3,reducedDim-method}
\alias{plotPath3,reducedDim-method}
\title{Plot path in 3D plot. Only to be called after \code{plot3}}
\usage{
\S4method{plotPath3}{reducedDim}(object, diamWidth, diamCol, useDims,
  selectedPath, project2D)
}
\arguments{
\item{object}{Dimensionality Reduced object}

\item{diamWidth}{Single length numeric controlling the diameter path line}

\item{diamCol}{Single value for color of the diameter path}

\item{useDims}{Dimensions to be used}

\item{selectedPath}{Name of path stored in paths slot}

\item{project2D}{Default NULL. Projects 3D graph into 2D by the given transformation
matrix. This matrix must have been produced using \code{getMVP} on an open rgl device.}
}
\value{
Plots 3D path
}
\description{
3D path plot
}
\details{
3D path plot
}
\author{
Wajid Jawaid
}
