% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotQC}
\alias{plotQC}
\title{Plots QC from SCD}
\usage{
plotQC(scd, catG = NULL, ylims = NULL)
}
\arguments{
\item{scd}{Single cell dataset}

\item{catG}{Assign colour to each cell. Named vector of colours with
names corressponding to cells.}

\item{ylims}{List object with a vector ylim (lower limit, upper limit)
in each entry of the list.}
}
\value{
Generates plot.
}
\description{
Plots QC from SCD
}
\details{
Plots QC from SCD
}
\author{
wj241
}

