% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotTechVar}
\alias{plotTechVar}
\title{Plots a technical noise analysis plot}
\usage{
plotTechVar(varGenes, pdf = NULL, pch = 20, pch.var = pch,
  pch.notvar = pch, pch.spike = pch, ...)
}
\arguments{
\item{varGenes}{Object returned from \code{techVar()} function}

\item{pdf}{Default NULL. If anything other than NULL will produce a pdf of that name.}

\item{pch}{Default 20. Point type to use in plot, if specific points are not defined.
See below. Parameter is plossed to base R \code{plot} function}

\item{pch.var}{Defaults to pch above if not defined. Point type for highly variable genes.}

\item{pch.notvar}{Defaults to pch above if not defined. Point type for not highly
variable genes.}

\item{pch.spike}{Defaults to pch above if not defined. Point type for spike-ins.}

\item{...}{Further parameters to pass to plot}
}
\value{
Produces a plot.
}
\description{
Plots a technical noise analysis plot
}
\details{
Plots a technical noise analysis plot
}
\author{
Wajid Jawaid
}
