% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plotTree3}
\alias{plotTree3}
\title{Plots given path}
\usage{
plotTree3(object, useDims, pointColor, ...)
}
\arguments{
\item{object}{Dimensionality reduced object}

\item{useDims}{Dimensions to be plotted}

\item{pointColor}{Color for each point}

\item{...}{Parameters passed to plot}
}
\value{
A path
}
\description{
2D/3D Path plotting function from package \code{bglab}
}
\details{
2D/3D Path plotting function from package \code{bglab}.
This function requires \code{plot3} to have been called
}
\author{
Wajid Jawaid
}
