% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepData}
\alias{prepData}
\title{Prepares data frame}
\usage{
prepData(eD, aD, nm)
}
\arguments{
\item{eD}{Annotation from gene expression matrix}

\item{aD}{Annotated data frame to be prepared.}

\item{nm}{Name to pass in case of error.}
}
\value{
Returns prepared data frame
}
\description{
Prepares annotated data frame
}
\details{
Prepares annotated data frame
}
\author{
Wajid Jawaid
}
