% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{printGene}
\alias{printGene}
\title{Produce pdf of \code{plotGene} output}
\usage{
printGene(gene, directory, data)
}
\arguments{
\item{gene}{Gene to be visualised.}

\item{directory}{Directory where to place file, without trailing forward slash.}

\item{data}{Gene expression data from \code{exprs} of single cell dataset object.}
}
\value{
Produces a file of the name "gene.pdf" in \code{directory}.
}
\description{
Produce pdf of \code{plotGene} output
}
\details{
Produce pdf of \code{plotGene} output
}
\author{
Wajid Jawaid
}
