% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{qc,SCD-method}
\alias{qc,SCD-method}
\title{Retrieve QC in counts}
\usage{
\S4method{qc}{SCD}(object)
}
\arguments{
\item{object}{SCD}
}
\value{
Matrix of QC counts
}
\description{
Retrieve QC in counts
}
\details{
Retrieve QC in counts
}
\author{
Wajid Jawaid
}
