% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{qc}
\alias{qc}
\title{Retrieve QC in counts}
\usage{
qc(object, ...)
}
\arguments{
\item{object}{SCD}

\item{...}{Additional parameters}
}
\value{
Matrix of QC counts
}
\description{
Retrieve QC counts
}
\details{
Retrieve QC counts
}
\author{
Wajid Jawaid
}
