% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qcFunc}
\alias{qcFunc}
\title{Quality control function}
\usage{
qcFunc(counts, ercc.data, htseqQC, geneTable, meta, metaLaneID = "flowCell",
  mitochondrialIdenitifier = "^mt|^MT", pdf = NULL, cutoffs = c(2e+05,
  0.2, 0.2, 0, 0, 1, 0), orientation = c("landscape", "portrait"))
}
\arguments{
\item{counts}{raw counts}

\item{ercc.data}{Spike in data}

\item{htseqQC}{HTseq output}

\item{geneTable}{Genetable}

\item{meta}{Metadata dataframe}

\item{metaLaneID}{Lane that contains flowcell lane information}

\item{mitochondrialIdenitifier}{Prefix for mitochondrial genes in geneTable}

\item{pdf}{Name of output pdf}

\item{cutoffs}{Default 2e5, 0.2, 0.2, 0, 0, 1, 0.
Set thresholds for:
\itemize{
 \item{number of mapped reads (greater than threshold), }
 \item{ratio of genes to total number of reads (greater than threshold), }
 \item{ratio of mitchondrial genes to mitochondrial + nuclear genes (less than threshold),}
 \item{number of nuclear genes (greater than threshold), }
 \item{number of genes with >10 reads per million (greater than threshold),}
 \item{ercc:mapped (less than threshold) and}
 \item{nuclearGenes:mapped (greater than threshold).}
}}

\item{orientation}{Default "landscape". Can also be "portrait".}
}
\value{
List
}
\description{
Quality control function
}
\details{
Quality control function
}
\author{
Wajid Jawaid
}
