% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_defs.R
\docType{class}
\name{reducedDim-class}
\alias{reducedDim-class}
\title{A virtual S4 class}
\description{
\code{reducedDim} is a virtual S4 class used by \linkS4class{PCA} and
\linkS4class{DiffusionMap} classes.
}
\section{Slots}{

\describe{
\item{\code{eigenvalues}}{A vector of eigen values}

\item{\code{eigenvectors}}{A matrix with eigen vectors in columns.}

\item{\code{graph}}{Contains the minimum spanning tree}

\item{\code{paths}}{List of the various calculated paths}
}}

