% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{removePath,SCD-method}
\alias{removePath,SCD-method}
\title{Removes paths from object}
\usage{
\S4method{removePath}{SCD}(object, reducedDim, paths)
}
\arguments{
\item{object}{A singel cell dataset object}

\item{reducedDim}{The reduced dimension object to use}

\item{paths}{Paths to be removed. Can be numeric or character vector of path names.
If "ALL" then all paths will be removed.}
}
\value{
SCD
}
\description{
Removes paths from object
}
\details{
Removes paths from object
}
\author{
Wajid Jawaid
}

