% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{removePath}
\alias{removePath}
\title{Generic function to remove saved paths}
\usage{
removePath(object, paths, ...)
}
\arguments{
\item{object}{A singel cell dataset object}

\item{paths}{Paths to be removed. Can be numeric or character vector of path names.
If "ALL" then all paths will be removed.}

\item{...}{Additional parameters}
}
\value{
A single cell dataset object
}
\description{
Generic function to remove saved paths
}
\details{
Generic function to remove saved paths
}
\author{
Wajid Jawaid
}
