% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{resetCells}
\alias{resetCells}
\title{Include all cells}
\usage{
resetCells(scd)
}
\arguments{
\item{scd}{SCD}
}
\value{
Updated SCD
}
\description{
Include all cells
}
\details{
Include all cells
}
\author{
wj241
}

