% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{runPCA,SCD-method}
\alias{runPCA,SCD-method}
\title{Principal Component Analysis}
\usage{
\S4method{runPCA}{SCD}(object, ...)
}
\arguments{
\item{object}{Single Cell Dataset object}

\item{...}{Parameters passed to \code{prcomp}}
}
\value{
Single Cell Dataset object
}
\description{
Calculate PCA for Single Cell Dataset
}
\details{
Calculate PCA for Single Cell Dataset
}
\author{
Wajid Jawaid
}
