% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{runTSNE}
\alias{runTSNE}
\title{Perform t-distributed stochastic neighbour embedding}
\usage{
runTSNE(object, ndims = 3, perplexity = 30, theta = 0.5, pca = FALSE,
  seed = NULL, verbose = FALSE, use_dist = FALSE, dist_fun = NULL, ...)
}
\arguments{
\item{object}{Single Cell Data Set object}

\item{ndims}{Number of dimensions for new embedding}

\item{perplexity}{Perplecity parameter}

\item{theta}{Theta parameter}

\item{pca}{Logical, Perform initial PCA reduction}

\item{seed}{Set random seed}

\item{verbose}{To show output}

\item{use_dist}{Default FALSE. Calculate a distance metric on the
data and use this in the Rtsne function.}

\item{dist_fun}{Default NULL. If NULL will use (1 - spearman_correlation(data)) / 2.}

\item{...}{Additional parameters to be passed to Rtsne}
}
\value{
Returns an upadate Single Cell Dataset object
}
\description{
Retrieve diffusion map object
}
\details{
Retrieve diffusion map object
}
\author{
Wajid Jawaid
}
