% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_defs.R
\docType{class}
\name{SCD-class}
\alias{SCD-class}
\title{Single Cell Dataset}
\description{
This is the main class in the \pkg{bglab} package. It stores all the data. Three datatables
are required to generate a new object. Both qPCR and RNAseq data may be used.
}
\section{Slots}{

\describe{
\item{\code{useFilter}}{Logical. To filter dataset or not}

\item{\code{filterGene}}{Logical. To filter genes from dataset or not}

\item{\code{filterCell}}{Logical. To filter cells from dataset or not}

\item{\code{pca}}{Contains a \linkS4class{PCA} object}

\item{\code{diffMap}}{Contains a \linkS4class{DiffusionMap} object}

\item{\code{tsne}}{Contains a \linkS4class{TSNE} object}

\item{\code{isomap}}{Contains a \linkS4class{Isomap} object}

\item{\code{counts}}{Contains a matrix of raw counts.}

\item{\code{spike}}{Spike-in matrix.}

\item{\code{qcCounts}}{Quality Control matrix.}

\item{\code{filterQC}}{Default TRUE. Will not return cells that failed QC when set to TRUE.}

\item{\code{normalise}}{Either "DESeq" or "scran".}

\item{\code{qcOutput}}{List containing QC output.}

\item{\code{technicalNoise}}{List containing technical noise output.
Highly recommended.}
}}

