% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{selectCells,SCD-method}
\alias{selectCells,SCD-method}
\title{Select cells function}
\usage{
\S4method{selectCells}{SCD}(object, reduceMethod, useDims = 1:3,
  mark = TRUE, returnInd = FALSE, label.cex = 0.3, pdf = NULL, ...)
}
\arguments{
\item{object}{Single Cell Dataset object}

\item{reduceMethod}{Dimensionality reduction method to choose. At the moment can be either "pca" or "diffMap"}

\item{useDims}{Dimensions to be plotted}

\item{mark}{Logical. Default "TRUE" marks the cells with names on the plot}

\item{returnInd}{Logical. Default "FALSE" returns a numeric vector of named indices}

\item{label.cex}{Default 0.3. Set for label size}

\item{pdf}{Defaut is NULL. Else will generate a pdf.}

\item{...}{Further parameters passed to the plot function}
}
\value{
Returns a plot with marked cells and a vector of either named indices or just
names if "returnId" is set to "FALSE"".
}
\description{
Used for selecting cells
}
\details{
Used for selecting cells. Generates plot and then allows user to choose cells. Cells
are marked on the plot and are returned.
}
\author{
Wajid Jawaid
}

