% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{selectionTray}
\alias{selectionTray}
\title{Selection Heatmap}
\usage{
selectionTray(scd, retHeat, pdf = FALSE, data = NULL)
}
\arguments{
\item{scd}{Single Cell Dataset object.}

\item{retHeat}{List returned from heatmap original heatmap.}

\item{pdf}{Will it be output to a pdf}

\item{data}{If the data to heatmap was altered, enter the new data as a matrix.}
}
\value{
Plots heatmap and returns the display.
}
\description{
Displays heatmap to choose area for submap
}
\details{
Displays heatmap to choose area for submap
}
\author{
Wajid Jawaid
}

