% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{selectVariableGenes,SCD-method}
\alias{selectVariableGenes,SCD-method}
\title{To select only highly variable genes for further analysis}
\usage{
\S4method{selectVariableGenes}{SCD}(object, includeGenes = NULL,
  excludeGenes = NULL, reset = FALSE)
}
\arguments{
\item{object}{Single Cell dataset object}

\item{includeGenes}{Genes to be included}

\item{excludeGenes}{Genes to be excluded}

\item{reset}{Default is FALSE. If TRUE will set selected genes as provided and all others
as the alternative. If FALSE will leave other genes unchanged.}
}
\value{
Returns modififed Single Cell Dataset object
}
\description{
To select only highly variable genes for further analysis
}
\details{
To select only highly variable genes for further analysis. Give either genes to
be included or genes to be excluded BY ENSMBL id only. Do not give both.
}
\author{
Wajid Jawaid
}
