% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{sf,SCD-method}
\alias{sf,SCD-method}
\title{Retrieve size factors}
\usage{
\S4method{sf}{SCD}(object, type = "bio")
}
\arguments{
\item{object}{SCD}

\item{type}{Default "bio". For "spike-ins" use "tech"}
}
\value{
Returns matrix of size factors
}
\description{
Retrieve size factors
}
\details{
Retrieve size factors
}
\author{
wj241
}

