% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{sf<-,SCD,character,matrix-method}
\alias{sf<-,SCD,character,matrix-method}
\title{Save the size factors}
\usage{
\S4method{sf}{SCD,character,matrix}(object, type) <- value
}
\arguments{
\item{object}{SCD}

\item{type}{Can be "bio" or "tech"}

\item{value}{Matrix of values}
}
\value{
UYpdated SCD object
}
\description{
Save the size factors
}
\details{
Save the size factors
}
\author{
wj241
}
