% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sf<-}
\alias{sf<-}
\title{Save the size factors}
\usage{
sf(object, type) <- value
}
\arguments{
\item{object}{SCD}

\item{type}{Can be "bio" or "tech"}

\item{value}{Matrix of values}
}
\value{
UYpdated SCD object
}
\description{
Save the size factors
}
\details{
Save the size factors
}
\author{
wj241
}
