% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sf}
\alias{sf}
\title{Retrieve size factors}
\usage{
sf(object, type = "bio")
}
\arguments{
\item{object}{SCD}

\item{type}{Default "bio". For "spike-ins" use "tech"}
}
\value{
Returns matrix of size factors
}
\description{
Retrieve size factors
}
\details{
Retrieve size factors
}
\author{
wj241
}

