% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{sizeFactors<-,SCD,numeric-method}
\alias{sizeFactors<-,SCD,numeric-method}
\title{Set size factors}
\usage{
\S4method{sizeFactors}{SCD,numeric}(object, type = NULL, ...) <- value
}
\arguments{
\item{object}{SCD Single Cell Dataset}

\item{type}{optional character argument providing the type or name of the
size factors to be accessed or assigned.}

\item{...}{Additional parameters}

\item{value}{A vector to be stored}
}
\value{
SCD object
}
\description{
Set size factors
}
\details{
Set size factors
}
\author{
Wajid Jawaid
}
