% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{slidingAv}
\alias{slidingAv}
\title{Calculates sliding average}
\usage{
slidingAv(data, winLength = 10)
}
\arguments{
\item{data}{Numeric vector for which sliding average is to be calculated}

\item{winLength}{Integer defining window length to be used}
}
\value{
Numeric vector of sliding average for the given window length
}
\description{
Calculates sliding average
}
\details{
Calculates sliding average
}
\author{
Wajid Jawaid
}
