% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{smoothGn}
\alias{smoothGn}
\title{Gam smoothing of gene expression}
\usage{
smoothGn(data, bs = "tp", family = gaussian(), n = length(data), ...)
}
\arguments{
\item{data}{Vector of data}

\item{bs}{Default "tp", represents thin plate, may be cubic or other.}

\item{family}{Deafult gaussian().}

\item{n}{Not used.}

\item{...}{Additional parameters to pass to predict}
}
\value{
Returns vector of smoothed values
}
\description{
Gam smoothing of gene expression
}
\details{
Gam smoothing of gene expression
}
\author{
Wajid Jawaid
}

