% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sparseMarkov}
\alias{sparseMarkov}
\title{Generate a sparse Markov matrix for momentum method to generate training data.
Utilises two approaches. First removes edges that are below a threshold for the probability if n number of cells are equally connected, i.e. if all cells are connected all will have an equal but low transition probability and this may be considered to be a disconnected node.
Second cells that are more highly connected are allowed to keep greater number of nearest neoghbours.}
\usage{
sparseMarkov(x, nn = ceiling(rowSums(x)), n = ncol(x) - 1)
}
\arguments{
\item{x}{Square matrix representing proximity. Try using Gaussian kernel from
maps.}

\item{nn}{Default ceiling(rowSums(x)). Number of nearest neighbours. The default tries
to ensure that highly connected nodes remain highly connected.}

\item{n}{Number of cells with equal binding threshold.}
}
\value{
Returns normalised transition matrix.
}
\description{
Generate a sparse Markov matrix for momentum method to generate training data
}
\details{
Generate a sparse Markov matrix for momentum method to generate training data
}
\author{
Wajid Jawaid
}
