% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{spikes,SCD-method}
\alias{spikes,SCD-method}
\title{Retrieve spike in counts}
\usage{
\S4method{spikes}{SCD}(object)
}
\arguments{
\item{object}{SCD}
}
\value{
Matrix of spike counts
}
\description{
Retrieve spike in counts
}
\details{
Retrieve spike in counts
}
\author{
Wajid Jawaid
}

