% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{spikes}
\alias{spikes}
\title{Retrieve spike in counts}
\usage{
spikes(object, ...)
}
\arguments{
\item{object}{SCD}

\item{...}{Additional parameters}
}
\value{
Matrix of spike counts
}
\description{
Retrieve spike in counts
}
\details{
Retrieve spike in counts
}
\author{
Wajid Jawaid
}

