% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{submap}
\alias{submap}
\title{Subset Heatmap.}
\usage{
submap(scd, retHeat, sTdev = dev.cur(), pdf = NULL, boxLwid = 3,
  boxCol = "black", data = NULL)
}
\arguments{
\item{scd}{Single Cell Dataset object.}

\item{retHeat}{List returned from heatmap original heatmap.}

\item{sTdev}{The display on which selectionTray is displayed}

\item{pdf}{Filename to export pdf}

\item{boxLwid}{Line width of box showing selection}

\item{boxCol}{Colour of box line}

\item{data}{If the data to heatmap was altered, enter the new data as a matrix.}
}
\value{
Produces a heatmap to select points and then a submap.
}
\description{
Generate heatmaps as subsets of a larger heatmap
}
\details{
Generate heatmaps as subsets of a larger heatmap. Call selectionTray
before calling this function and select top.
left and bottom right corners to chose selection.
}
\author{
Wajid Jawaid
}
