% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{techVar,SCD-method}
\alias{techVar,SCD-method}
\title{Technical noise analysis for single cell RNAseq}
\usage{
\S4method{techVar}{SCD}(object, useERCC = TRUE, cvThresh = 0.3,
  quant = 0.8, minBiolDisp = 0.5^2, fdr = 0.1, data.lengths = NULL,
  ercc.lengths = NULL, meanForFit = NULL)
}
\arguments{
\item{object}{SCD class}

\item{useERCC}{Default TRUE. To use ERCC for technical noise estimation.}

\item{cvThresh}{Defalt 0.3. The CV threshold to use for choosing cells to fit. See quant.}

\item{quant}{Default 0.8. The quantile of data to fit so that linear model is not fitted to
plateau at lower gene expression values. See Breenecke et al. for further details}

\item{minBiolDisp}{Default 0.25. Assumed biological dispersion. See paper.}

\item{fdr}{Deafult .1. False Discovery Rate for chi-squared test.}

\item{data.lengths}{Gene lengths in kilobases for PK "per kilobase normalisation"}

\item{ercc.lengths}{ERCC lengths in kilobases for PK "per kilobase normalisation"}

\item{meanForFit}{Provide a minimum mean for fit. (OPTIONAL)}
}
\value{
Returns updated SCD
}
\description{
Normalises and performs technical noise analysis for single cell RNAseq(Brennecke et al.)
}
\details{
Normalises and performs a technical noise analysis for single cell RNAseq.
Give raw counts. Data is normalised using size-factor normalisation identical to DESeq2,
then a technical noise analysis is perfromed as described in Brennecke et al. Normalised
expression values are log10 normalised.
}
\author{
Wajid Jawaid
}
