% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tp}
\alias{tp}
\title{Calculate cubeless, gammonless take points at different scores}
\usage{
tp(a, b, cube, met, last_roll = FALSE)
}
\arguments{
\item{a}{number of points that player needs}

\item{b}{number of points that opponent needs}

\item{cube}{cube value (before doubling)}

\item{met}{match equity table}

\item{last_roll}{treat this as a last roll position; no automatic redouble available}
}
\value{
double. Take point
}
\description{
Calculate cubeless, gammonless take points at different scores
}
\examples{
met <- get_met()
tp(3, 5, 1, met)

}
