% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tp_gammons}
\alias{tp_gammons}
\title{Calculate cubeless, take points at different scores, including gammons and backgammons}
\usage{
tp_gammons(a, b, gamfreq_a, bgfreq_a, gamfreq_b, bgfreq_b, cube, met)
}
\arguments{
\item{a}{number of points that player needs}

\item{b}{number of points that opponent needs}

\item{gamfreq_a}{proportion of player's wins that are gammons}

\item{bgfreq_a}{proportion of player's wins that are backgammons}

\item{gamfreq_b}{proportion of player's losses that are gammons}

\item{bgfreq_b}{proportion of player's losses that are backgammons}

\item{cube}{cube value (before doubling)}

\item{met}{match equity table}
}
\value{
double. Take point
}
\description{
Calculate cubeless, take points at different scores, including gammons and backgammons
}
