% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tp_janowski}
\alias{tp_janowski}
\title{Calculate cubeful take points for money game, Janowski-style}
\usage{
tp_janowski(W, L, x = 2/3)
}
\arguments{
\item{W}{Average cubeless value of games ultimately won}

\item{L}{Average cubeless value of games ultimately lost}
}
\value{
double. Take point
}
\description{
Calculate cubeful take points for money game, Janowski-style
}
