% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tp_table}
\alias{tp_table}
\title{Compute take points at a given match score, create a nice table
to compare those with money game take points}
\usage{
tp_table(x, y, probs, cube, met)
}
\arguments{
\item{x}{number of points that player needs}

\item{y}{number of points that opponent needs}

\item{probs}{numeric vector of length 6, representing outcome
probabilities (must sum to 1 or 100)}

\item{cube}{cube value}

\item{met}{match equity table}
}
\value{
data.frame
}
\description{
Compute take points at a given match score, create a nice table
to compare those with money game take points
}
\examples{
met <- get_met()
# XGID=-a-BaBC-A---eE---c-e----B-:0:0:1:00:0:0:0:0:10
# 4-ply winning chances, reported in a cumulative fashion:
cum_probs <- c(61.94, 24.09, 1.04, 38.06, 8.54, 0.42)
probs <- outcome_probs(cum_probs)
tp_table(3, 5, probs, 1, met)

}
