% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tp_xg}
\alias{tp_xg}
\title{Take points according to eXtreme Gammon}
\usage{
tp_xg(x, y, probs, cube, met, last_roll = FALSE)
}
\arguments{
\item{x}{number of points that player needs}

\item{y}{number of points that opponent needs}

\item{probs}{numeric vector of length 6, representing outcome
probabilities (must always sum to 1 or 100)}

\item{cube}{cube value (before doubling)}

\item{met}{match equity table}

\item{last_roll}{this as a last roll position; no automatic redouble available.
Defaults to FALSE}
}
\value{
List of take point: 'Dead Cube' and 'Live Cube'
}
\description{
Calculate dead cube and live cube take points like eXtreme Gammon does
in the 'Cube Information' window. Based on match score score, gammons,
and backgammons, cube level, and match equity table
}
