% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_defs.R
\docType{class}
\name{TSNE-class}
\alias{TSNE-class}
\title{TSNE dimensionality reduction on dataset}
\description{
Performed in Matlab - Barnes-Hut fast tSNE algorithm used
}
\section{Slots}{

\describe{
\item{\code{theta}}{Theta parameter}

\item{\code{perplexity}}{Perplexity parameter}

\item{\code{N}}{Number of cells/samples}

\item{\code{origD}}{Original number of cells/samples}

\item{\code{seed}}{Random seed used, saved for reproducibility}

\item{\code{nnError}}{The tSNE nearest neighbour error score}

\item{\code{algorithm}}{The algorithm used e.g. Barnes-Hut}
}}

