% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{txt2df}
\alias{txt2df}
\title{Convert analysis files to data frame}
\usage{
txt2df(files)
}
\arguments{
\item{files}{character vector with names of *.txt files to parse}
}
\value{
tibble
}
\description{
Parse GNU Backgammon Galaxy analysis *.txt files into data frame format
}
