% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{useFilter,SCD-method}
\alias{useFilter,SCD-method}
\title{Predicate indicating whether overall filtering is active}
\usage{
\S4method{useFilter}{SCD}(object)
}
\arguments{
\item{object}{Single Cell Dataset}
}
\value{
Logical.
}
\description{
To use filter or not
}
\details{
To use filter or not.
}
\author{
Wajid Jawaid
}
