% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{useFilter<-,SCD,logical-method}
\alias{useFilter<-,SCD,logical-method}
\title{Set whether to use filter or not}
\usage{
\S4method{useFilter}{SCD,logical}(object) <- value
}
\arguments{
\item{object}{SCD object}

\item{value}{Logical value.}
}
\value{
Updated SCD object
}
\description{
Set whether to use filter or not
}
\details{
Set whether to use filter or not
}
\author{
Wajid Jawaid
}

