% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{useFilter}
\alias{useFilter}
\title{Add additional phenotype data}
\usage{
useFilter(object)
}
\arguments{
\item{object}{Single Cell Dataset}
}
\value{
Logical.
}
\description{
To use filter or not
}
\details{
To use filter or not. Replaces "noprocessing" argument from previous version
}
\author{
Wajid Jawaid
}
