% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{winsor}
\alias{winsor}
\title{Winsorization function}
\usage{
winsor(x, fraction = 0.05, lower = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{fraction}{The fraction of cells on}

\item{lower}{By default this function will only winsorise the upper end.
If you also require the bottom end to be winsorised set this to TRUE.}
}
\value{
Return winsorized vector
}
\description{
Winsorization function
}
\details{
Winsorization function.
Transformation limiting spurious extreme values.
Will set all outliers to a specific percentile of the data.
}
\author{
Wajid Jawaid
}
