% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{writeGeneList}
\alias{writeGeneList}
\title{Write a list of gene vectors to file}
\usage{
writeGeneList(listofGenes, file)
}
\arguments{
\item{listofGenes}{Named list containing vectors of genes. Names
will be used as a header}

\item{file}{Filename for export}
}
\value{
Returns data.frame of list with coloumn post-column binding.
}
\description{
Write a list of gene vectors to file
}
\details{
Write a list of gene vectors to file
}
\author{
Wajid Jawaid
}

