% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xgid2vec}
\alias{xgid2vec}
\title{Checker locations as integer vector}
\usage{
xgid2vec(xgid)
}
\arguments{
\item{xgid}{}
}
\value{
vector of length 26 with signed integers
}
\description{
Takes the checker locations from a XGID string (first 26 elements)
and turns it into a signed integer vector of length 26, where
positive numbers indicate how many checkers one player has on a
point; negative numbers are the other player's count of checkers.
}
\examples{
xgid2vec("XGID=-b----E-C---eE---c-e----B-:0:0:1:52:0:0:3:0:10")
}
