% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{emg}
\alias{emg}
\title{Scale match winning chance to equivalent to money game}
\usage{
emg(mwc, a, b, cube, met)
}
\arguments{
\item{mwc}{current match winning chances for player}

\item{a}{number of points that player needs}

\item{b}{number of points that opponent needs}

\item{cube}{cube value}

\item{met}{match equity table}
}
\value{
double
}
\description{
Scale match winning chance to equivalent to money game
}
