% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_met}
\alias{get_met}
\title{Get match equity table from *.met file (used by Extreme Gammon)}
\usage{
get_met(filename = "data-raw\\\\Kazaross XG2.met")
}
\arguments{
\item{filename}{file location}
}
\value{
matrix
}
\description{
Get match equity table from *.met file (used by Extreme Gammon)
}
