% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggboard.R
\name{ggboard}
\alias{ggboard}
\title{Display a backgammon diagram from an eXtreme Gammon id}
\usage{
ggboard(xgid)
}
\arguments{
\item{xgid}{character}
}
\value{
ggplot object
}
\description{
Display a backgammon diagram from an eXtreme Gammon id
}
