% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posid2xgid.R
\name{gnuid2xgid}
\alias{gnuid2xgid}
\title{From GNU BG ids to XGIDs}
\usage{
gnuid2xgid(pos_id, match_id)
}
\arguments{
\item{pos_id}{character. GNU Backgammon position id}

\item{match_id}{character. GNU Backgammon match id}
}
\value{
character
}
\description{
Convert GNU Backgammon position and match ids to XGID, the format
used by eXtreme Gammon
}
\examples{
gnuid2xgid(pos_id = "4HPwATDgc/ABMA", match_id = "8IhuACAACAAE")

}
