% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{newSCD}
\alias{newSCD}
\title{Make a new SCD object}
\usage{
newSCD(experimentType, assayData = NULL, featureData, phenoData, counts,
  spike = NULL, qc = NULL, cellPairwiseDistances = dist(vector()),
  featurePairwiseDistances = dist(vector()), ...)
}
\arguments{
\item{experimentType}{Experiment Type. RNAseq or qPCR.}

\item{assayData}{Matrix of data with genes in rows and cells in columns}

\item{featureData}{Data frame of further details for each gene. Rownames must
match row names of assayData.}

\item{phenoData}{Data frame of further details for each cell. Rownames must
match column names of assayData.}

\item{counts}{Raw counts data.}

\item{spike}{Default NULL. Include spike-in data e.g. ERCC-92.}

\item{qc}{Default NULL. Include quality control counts from HTseq for example.}

\item{cellPairwiseDistances}{CPW}

\item{featurePairwiseDistances}{FPW}

\item{...}{Further parameters.}
}
\value{
New Single Cell Dataset object.
}
\description{
Make new SCD object
}
\details{
Construct a new Single cell dataset object for the bglab package
}
\author{
Wajid Jawaid
}
