% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{probs_table}
\alias{probs_table}
\title{Probability table}
\usage{
probs_table(probs, margins = TRUE)
}
\arguments{
\item{probs}{numeric vector of length 6, representing outcome
probabilities (must sum to 1 or 100)}

\item{margins}{boolean. Should margin totals be added?
Defaults to TRUE}
}
\value{
data.frame with rownames
}
\description{
Put winning probabilities in nice table
}
\examples{
# XGID=-a-BaBC-A---eE---c-e----B-:0:0:1:00:0:0:0:0:10
# 4-ply winning chances, reported in a cumulative fashion:
cum_probs <- c(61.94, 24.09, 1.04, 38.06, 8.54, 0.42)
probs <- outcome_probs(cum_probs)
probs_table(probs)

}
