% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tp}
\alias{tp}
\title{Simple take point calculation}
\usage{
tp(x, y, cube, met, last_roll = FALSE)
}
\arguments{
\item{x}{number of points that player needs}

\item{y}{number of points that opponent needs}

\item{cube}{cube value (before doubling)}

\item{met}{match equity table}

\item{last_roll}{treat this as a last roll position; no automatic redouble available}
}
\value{
double. Take point
}
\description{
Calculate cubeless, gammonless take points at different scores
}
\examples{
met <- get_met() # Get the default Kazaross XG2 table
tp(3, 5, 1, met)

}
