% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tp_gammons}
\alias{tp_gammons}
\title{Take point calculation including gammons}
\usage{
tp_gammons(x, y, probs, cube, met)
}
\arguments{
\item{x}{number of points that player needs}

\item{y}{number of points that opponent needs}

\item{probs}{numeric vector of length 6, representing outcome
probabilities (must always sum to 1 or 100)}

\item{cube}{cube value (before doubling)}

\item{met}{match equity table}
}
\value{
double. Take point
}
\description{
Calculate cubeless take points at different scores, including gammons and backgammons
}
