% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tp_janowski}
\alias{tp_janowski}
\title{Calculate take points for money game, Janowski-style}
\usage{
tp_janowski(probs, x = 2/3)
}
\arguments{
\item{probs}{numeric vector of length 6, representing outcome
probabilities (must always sum to 1 or 100)}

\item{x}{cube-life index, between 0 and 1}
}
\value{
double. Take point
}
\description{
Calculate take points for money game, Janowski-style
}
