% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tp_money}
\alias{tp_money}
\title{Take points, money game}
\usage{
tp_money(probs, cube_eff = 2/3)
}
\arguments{
\item{probs}{numeric vector of length 6, representing outcome
probabilities (must always sum to 1 or 100)}

\item{cube_eff}{cube-life index, between 0 and 1, defaults to 2/3}
}
\value{
double. Take point
}
\description{
Calculate take points for money game, Janowski-style
}
